refactor: merge diecut-core and diecut-cli into single crate
Eliminates the Cargo workspace by combining the library and CLI into
one package. The project only ships a binary; the two-crate split
added complexity (especially for release-please) without benefit.

Also simplifies release-please config to a single "." package with
no plugins needed.
