+++ .release-please-manifest.json
@@ -1,4 +1,3 @@
 {
-  "crates/diecut-cli": "0.1.1",
-  "crates/diecut-core": "0.1.1"
+  ".": "0.1.0"
 }
