--- crates/diecut-cli/CHANGELOG.md
@@ -1,26 +1,0 @@
-# Changelog
-
-## [0.1.1](https://github.com/raiderrobert/diecut/compare/diecut-v0.1.0...diecut-v0.1.1) (2026-02-13)
-
-
-### Features
-
-* support in-place migration with automatic backup ([e5f0de7](https://github.com/raiderrobert/diecut/commit/e5f0de7f0beab88d5a68ea4c608ec1f3c9235bb5))
-* support in-place migration with automatic backup ([4b586b8](https://github.com/raiderrobert/diecut/commit/4b586b8542a0df9ce34d7875527f57e3d76e3dbf))
-
-
-### Bug Fixes
-
-* configure release-please for Rust workspace per upstream pattern ([77c43c8](https://github.com/raiderrobert/diecut/commit/77c43c88fb16873642aef3cce012a25aeeae572f))
-* configure release-please for Rust workspace per upstream pattern ([4b61088](https://github.com/raiderrobert/diecut/commit/4b6108872df07a3bc7d52ae59637ff41a66d64c6))
-* remove cargo-workspace plugin and restore version.workspace ([1d074c9](https://github.com/raiderrobert/diecut/commit/1d074c9b3c8529f0ab8f7660f01617fac69df97e))
-* remove cargo-workspace plugin and restore version.workspace ([5d1c7c6](https://github.com/raiderrobert/diecut/commit/5d1c7c609d1ceb7ccd09f1a5e9cb7993876f8b80))
-* use explicit versions in subcrate Cargo.toml for release-please ([52df58f](https://github.com/raiderrobert/diecut/commit/52df58f60750b1df9fee0dc323cf56a69999d2a8))
-* use explicit versions in subcrate Cargo.toml for release-please ([e975299](https://github.com/raiderrobert/diecut/commit/e97529959f8ceee0179c015d1ffb0156d88e50ab))
-* use rename-swap for in-place migration instead of clear-and-copy ([c061f60](https://github.com/raiderrobert/diecut/commit/c061f609be8df344215a25814a82d3d00c2c7410))
-
-
-### Miscellaneous
-
-* remove verbose comments and simplify code ([027bb15](https://github.com/raiderrobert/diecut/commit/027bb158ed5ad48cefff362c622d50b4da4fada3))
-* remove verbose comments and simplify code across crates ([e5c5da9](https://github.com/raiderrobert/diecut/commit/e5c5da968f44b0becad28c739469b28f55310dcc))
