--- crates/diecut-cli/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "diecut"
-version = "0.1.1"
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "A language-agnostic project template generator"
-
-[[bin]]
-name = "diecut"
-path = "src/main.rs"
-
-[dependencies]
-diecut-core = { workspace = true }
-clap = { workspace = true }
-miette = { workspace = true }
-console = { workspace = true }
-tempfile = { workspace = true }
