--- crates/diecut-core/CHANGELOG.md
@@ -1,43 +1,0 @@
-# Changelog
-
-## [0.1.1](https://github.com/raiderrobert/diecut/compare/diecut-core-v0.1.0...diecut-core-v0.1.1) (2026-02-13)
-
-
-### Features
-
-* add diff3-style three-way conflict output in .rej files ([39e87f8](https://github.com/raiderrobert/diecut/commit/39e87f8a75ced98e9e7ff3f3e815e57764b2ffd9))
-* diff3-style three-way conflict output in .rej files ([4fd4e3b](https://github.com/raiderrobert/diecut/commit/4fd4e3b4f8f756931448487a87cc76fd1e8c82d2))
-
-
-### Bug Fixes
-
-* configure release-please for Rust workspace per upstream pattern ([77c43c8](https://github.com/raiderrobert/diecut/commit/77c43c88fb16873642aef3cce012a25aeeae572f))
-* configure release-please for Rust workspace per upstream pattern ([4b61088](https://github.com/raiderrobert/diecut/commit/4b6108872df07a3bc7d52ae59637ff41a66d64c6))
-* include file path in template render error messages ([5d87ddd](https://github.com/raiderrobert/diecut/commit/5d87ddd7f49fc42a9af942bb0fb8a919c22a7e1c))
-* include file path in template render error messages ([761651b](https://github.com/raiderrobert/diecut/commit/761651beca76d2c55d1daaf4afeb9790eaac828b))
-* read only first 8KB for binary file detection ([9fc4ed8](https://github.com/raiderrobert/diecut/commit/9fc4ed8253301c32d981f3ccf7ae6821564e4864))
-* read only first 8KB for binary file detection ([f52548a](https://github.com/raiderrobert/diecut/commit/f52548a3214d524bc4bcae6e3da35809dd78842c))
-* remove cargo-workspace plugin and restore version.workspace ([1d074c9](https://github.com/raiderrobert/diecut/commit/1d074c9b3c8529f0ab8f7660f01617fac69df97e))
-* remove cargo-workspace plugin and restore version.workspace ([5d1c7c6](https://github.com/raiderrobert/diecut/commit/5d1c7c609d1ceb7ccd09f1a5e9cb7993876f8b80))
-* sandbox Rhai hook engine to prevent filesystem access ([d2092e2](https://github.com/raiderrobert/diecut/commit/d2092e21dd6884eefcf178967e4793f6057ec8ba))
-* sandbox Rhai hook engine to prevent filesystem access ([6b2251e](https://github.com/raiderrobert/diecut/commit/6b2251ea38b2dbe7c8ad2f1ddbf0047c49bea8a5))
-* use content_inspector for BOM-aware binary detection ([d995178](https://github.com/raiderrobert/diecut/commit/d9951780a642e68edf3602cde586811268dba832))
-* use explicit versions in subcrate Cargo.toml for release-please ([52df58f](https://github.com/raiderrobert/diecut/commit/52df58f60750b1df9fee0dc323cf56a69999d2a8))
-* use explicit versions in subcrate Cargo.toml for release-please ([e975299](https://github.com/raiderrobert/diecut/commit/e97529959f8ceee0179c015d1ffb0156d88e50ab))
-* use OS-level advisory locks (fs4) for cache concurrency ([072072c](https://github.com/raiderrobert/diecut/commit/072072c9728c795865c2b68f4d4d187bbf9f03cc))
-* use OS-level advisory locks (fs4) for cache concurrency ([a0e3580](https://github.com/raiderrobert/diecut/commit/a0e3580f6ad13984e88a38579882630938b1358c))
-
-
-### Miscellaneous
-
-* remove unused indicatif dependency ([3486a78](https://github.com/raiderrobert/diecut/commit/3486a78908e1a336adc576335e5f2e372064265b))
-* remove unused indicatif dependency ([b19de48](https://github.com/raiderrobert/diecut/commit/b19de48d891200438c8f8504f6ce30348c6aee34))
-* remove verbose comments and simplify code ([027bb15](https://github.com/raiderrobert/diecut/commit/027bb158ed5ad48cefff362c622d50b4da4fada3))
-* remove verbose comments and simplify code across crates ([e5c5da9](https://github.com/raiderrobert/diecut/commit/e5c5da968f44b0becad28c739469b28f55310dcc))
-
-
-### Code Refactoring
-
-* extract write_cache_metadata and place_in_cache helpers ([8907726](https://github.com/raiderrobert/diecut/commit/8907726c21a73766e565132c6fe69eaeea493599))
-* extract write_cache_metadata and place_in_cache helpers ([1b9756a](https://github.com/raiderrobert/diecut/commit/1b9756a7edd7c31366aa6372dc26deebb4f18101))
-* use Option&lt;&str&gt; for commit_sha parameter ([2178c3b](https://github.com/raiderrobert/diecut/commit/2178c3b985bc54c4fe4d2314279227f425a8e293))
