--- crates/diecut-core/Cargo.toml
@@ -1,28 +1,0 @@
-[package]
-name = "diecut-core"
-version = "0.1.1"
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-description = "Core library for diecut template engine"
-
-[dependencies]
-tera = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-toml = { workspace = true }
-globset = { workspace = true }
-regex-lite = { workspace = true }
-thiserror = { workspace = true }
-miette = { workspace = true }
-walkdir = { workspace = true }
-inquire = { workspace = true }
-console = { workspace = true }
-rhai = { workspace = true }
-dirs = { workspace = true }
-gix = { workspace = true }
-tempfile = { workspace = true }
-sha2 = { workspace = true }
-similar = { workspace = true }
-fs4 = { workspace = true }
-content_inspector = { workspace = true }
