+++ src/adapter/cookiecutter.rs
