+++ src/adapter/detect.rs
