+++ src/adapter/mod.rs
