+++ src/adapter/native.rs
