+++ src/template/clone.rs
